+++ app/models/SensorModel.scala
@@ -1,7 +1,6 @@
 package models
 
 import play.api.libs.json._
-import play.api.libs.json.Json._
 import play.api.libs.functional.syntax._
 
 /**